Improve Apollo queries
