+++ assets/src/domain/eventEditor/services/apollo/queries/datetimes/useDatetimeIds.ts
@@ -4,7 +4,9 @@ import { EntityId } from '@appServices/apollo/types';
 const useDatetimeIds = (): EntityId[] => {
 	const datetimes = useDatetimes();
 
-	return datetimes.map(({ id }) => id);
+	// Sort the IDs list which may be out of order,
+	// thus changing the key used to access apollo cache
+	return datetimes.map(({ id }) => id).sort();
 };
 
 export default useDatetimeIds;
