+++ assets/src/domain/eventEditor/services/apollo/queries/index.ts
@@ -1,29 +1,11 @@
-import { GET_DATETIME, GET_DATETIMES } from './datetimes';
-import { GET_TICKET, GET_TICKETS } from './tickets';
-import { GET_PRICE, GET_PRICES } from './prices';
-import { GET_PRICE_TYPE, GET_PRICE_TYPES } from './priceTypes';
-import { GET_CURRENT_USER, GET_GENERAL_SETTINGS } from '../../../../shared/services/apollo/queries';
+// import { GET_DATETIME, GET_DATETIMES } from './datetimes';
+// import { GET_TICKET, GET_TICKETS } from './tickets';
+// import { GET_PRICE, GET_PRICES } from './prices';
+// import { GET_PRICE_TYPE, GET_PRICE_TYPES } from './priceTypes';
+// import { GET_CURRENT_USER, GET_GENERAL_SETTINGS } from '../../../../shared/services/apollo/queries';
 
 import { DatetimeEdge, TicketEdge, PriceEdge, PriceTypeEdge } from '../types';
 import { EntityEdge } from '@appServices/apollo/types';
-
-export const queries = {
-	/* datetimes */
-	GET_DATETIME,
-	GET_DATETIMES,
-	/* tickets */
-	GET_TICKET,
-	GET_TICKETS,
-	/* prices */
-	GET_PRICE,
-	GET_PRICES,
-	/* prices */
-	GET_PRICE_TYPE,
-	GET_PRICE_TYPES,
-	/* config */
-	GET_CURRENT_USER,
-	GET_GENERAL_SETTINGS,
-};
 
 export const DEFAULT_ENTITY_LIST_DATA: EntityEdge = {
 	nodes: [],
@@ -51,6 +33,8 @@ export const DEFAULT_PRICE_TYPE_LIST_DATA: PriceTypeEdge = {
 };
 
 export * from './datetimes';
+
+export * from './events';
 
 export * from './tickets';
