+++ assets/src/domain/eventEditor/services/apollo/queries/tickets/useTicketIds.ts
@@ -4,7 +4,9 @@ import { EntityId } from '@appServices/apollo/types';
 const useTicketIds = (): EntityId[] => {
 	const tickets = useTickets();
 
-	return tickets.map(({ id }) => id);
+	// Sort the IDs list which may be out of order,
+	// thus changing the key used to access apollo cache
+	return tickets.map(({ id }) => id).sort();
 };
 
 export default useTicketIds;
